Rename shared waf utils module
